adjust print format for size_t

+++ apps/sms/msmsdetailview.cc
@@ -281,7 +281,7 @@ static void mSmsDetailView_reloadData(mSmsDetailView* self)
         day_it->format(title, 64, "%Y-%m-%d");
         title_list.push_back(title);
         
-        printf("======================== tmp_sms_list size: %u %s\n", tmp_sms_list.size(), title);
+        printf("======================== tmp_sms_list size: %lu %s\n", (unsigned long)tmp_sms_list.size(), title);
     }
 
     releaseIcons(self);
